Refactors inventory management and adds tax calculations
Updates the inventory management forms to include tax and total calculations based on item quantity and price.
Refines the table presentation for inventory entries by leveraging a dedicated table resource to enhance organization.
Implements a unique code field for warehouses and ensures that code is enforced to uppercase on input.
Adds logic to handle document statuses and adjustments after record creation.

Relates to improvements in handling inventory operations effectively.

+++ database/migrations/2024_12_20_210136_create_documents_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->bigInteger('tax')->default(0);
             $table->bigInteger('total')->default(0);
             $table->tinyInteger('operation')->default(0); // 0: Sin efecto, 1: Entrada, 2: Salida
-            $table->unsignedBigInteger('status')->default(0);
+            $table->integer('status')->default(0);
             $table->uuid('uuid')->nullable()->unique();
             $table->timestamps();
